requirements-auth.txt
# 認証機能専用軽量requirements.txt

# 基本ライブラリ
streamlit>=1.32.0
python-dotenv>=1.0.0

# OAuth認証関連
authlib>=1.2.0
requests-oauthlib>=1.3.0
PyJWT>=2.8.0
cryptography>=41.0.0
streamlit-oauth>=0.0.1

# 基本的なHTTP通信
requests>=2.31.0

# AWS S3関連
boto3>=1.34.0

# 最小限のデータ処理（必要に応じて）
pandas>=2.2.0
numpy>=1.24.0

# 可視化ライブラリ
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
networkx>=3.1

# 日本語表示（必要に応じて）
japanize-matplotlib>=1.1.3

setuptools>=65.0.0